feat(workflow): add types for workflow definitions, tasks, fields, and artifacts
