+++ client/src/environments/environment.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: false,
-  apiUrl: 'http://localhost:3000/api/v2',
-  graphqlUrl: 'http://localhost:3000/api/v2/graphql'
+  apiUrl: 'http://localhost:3000/api',
+  graphqlUrl: 'http://localhost:3000/api/graphql'
 };
