+++ server/src/db/relations.ts
@@ -1,5 +1,5 @@
 import { relations } from "drizzle-orm/relations";
-import { notifications, notificationRecipients, users, faculty, departments, phaseSteps, programmePhaseSteps, programmePhases, attachments, phases, programmes, usersInWorkflow, programmesInWorkflow, facultyInWorkflow, departmentsInWorkflow, processInstancesInWorkflow, artifactsInWorkflow, taskInstancesInWorkflow, definitionsInWorkflow, definitionVersionsInWorkflow, auditEventsInWorkflow, phasesInWorkflow, phaseStepsInWorkflow, programmePhasesInWorkflow, programmePhaseStepsInWorkflow, attachmentsInWorkflow, notificationsInWorkflow, notificationRecipientsInWorkflow } from "./schema";
+import { notifications, notificationRecipients, users, faculty, departments, programmes, attachmentsInWorkflow, processInstancesInWorkflow, taskInstancesInWorkflow, definitionsInWorkflow, definitionVersionsInWorkflow, auditEventsInWorkflow } from "./schema";
 
 export const notificationRecipientsRelations = relations(notificationRecipients, ({one}) => ({
 	notification: one(notifications, {
@@ -18,8 +18,13 @@ export const notificationsRelations = relations(notifications, ({many}) => ({
 
 export const usersRelations = relations(users, ({many}) => ({
 	notificationRecipients: many(notificationRecipients),
-	attachments: many(attachments),
 	programmes: many(programmes),
+	attachmentsInWorkflows: many(attachmentsInWorkflow),
+	definitionVersionsInWorkflows: many(definitionVersionsInWorkflow),
+	processInstancesInWorkflows: many(processInstancesInWorkflow),
+	taskInstancesInWorkflows: many(taskInstancesInWorkflow),
+	definitionsInWorkflows: many(definitionsInWorkflow),
+	auditEventsInWorkflows: many(auditEventsInWorkflow),
 }));
 
 export const departmentsRelations = relations(departments, ({one}) => ({
@@ -33,136 +38,56 @@ export const facultyRelations = relations(faculty, ({many}) => ({
 	departments: many(departments),
 }));
 
-export const programmePhaseStepsRelations = relations(programmePhaseSteps, ({one, many}) => ({
-	phaseStep: one(phaseSteps, {
-		fields: [programmePhaseSteps.phaseStepId],
-		references: [phaseSteps.id]
-	}),
-	programmePhase: one(programmePhases, {
-		fields: [programmePhaseSteps.programmePhaseId],
-		references: [programmePhases.id]
-	}),
-	attachments: many(attachments),
-}));
-
-export const phaseStepsRelations = relations(phaseSteps, ({one, many}) => ({
-	programmePhaseSteps: many(programmePhaseSteps),
-	phase: one(phases, {
-		fields: [phaseSteps.phaseId],
-		references: [phases.id]
-	}),
-}));
-
-export const programmePhasesRelations = relations(programmePhases, ({one, many}) => ({
-	programmePhaseSteps: many(programmePhaseSteps),
-	programme: one(programmes, {
-		fields: [programmePhases.programmeId],
-		references: [programmes.id]
-	}),
-	phase: one(phases, {
-		fields: [programmePhases.phaseId],
-		references: [phases.id]
-	}),
-}));
-
-export const attachmentsRelations = relations(attachments, ({one}) => ({
-	programmePhaseStep: one(programmePhaseSteps, {
-		fields: [attachments.programmePhaseStepId],
-		references: [programmePhaseSteps.id]
-	}),
-	user: one(users, {
-		fields: [attachments.uploadedBy],
-		references: [users.id]
-	}),
-}));
-
-export const phasesRelations = relations(phases, ({many}) => ({
-	phaseSteps: many(phaseSteps),
-	programmePhases: many(programmePhases),
-}));
-
 export const programmesRelations = relations(programmes, ({one, many}) => ({
-	programmePhases: many(programmePhases),
 	user: one(users, {
 		fields: [programmes.initiator],
 		references: [users.id]
 	}),
-}));
-
-export const programmesInWorkflowRelations = relations(programmesInWorkflow, ({one, many}) => ({
-	usersInWorkflow: one(usersInWorkflow, {
-		fields: [programmesInWorkflow.initiator],
-		references: [usersInWorkflow.id]
-	}),
-	artifactsInWorkflows: many(artifactsInWorkflow),
-	processInstancesInWorkflows: many(processInstancesInWorkflow),
-	taskInstancesInWorkflows: many(taskInstancesInWorkflow),
-	auditEventsInWorkflows: many(auditEventsInWorkflow),
-	programmePhasesInWorkflows: many(programmePhasesInWorkflow),
-}));
-
-export const usersInWorkflowRelations = relations(usersInWorkflow, ({many}) => ({
-	programmesInWorkflows: many(programmesInWorkflow),
-	artifactsInWorkflows: many(artifactsInWorkflow),
-	definitionsInWorkflows: many(definitionsInWorkflow),
-	definitionVersionsInWorkflows: many(definitionVersionsInWorkflow),
-	processInstancesInWorkflows: many(processInstancesInWorkflow),
-	taskInstancesInWorkflows: many(taskInstancesInWorkflow),
-	auditEventsInWorkflows: many(auditEventsInWorkflow),
 	attachmentsInWorkflows: many(attachmentsInWorkflow),
-	notificationRecipientsInWorkflows: many(notificationRecipientsInWorkflow),
+	processInstancesInWorkflows: many(processInstancesInWorkflow),
+	taskInstancesInWorkflows: many(taskInstancesInWorkflow),
+	auditEventsInWorkflows: many(auditEventsInWorkflow),
 }));
 
-export const departmentsInWorkflowRelations = relations(departmentsInWorkflow, ({one}) => ({
-	facultyInWorkflow: one(facultyInWorkflow, {
-		fields: [departmentsInWorkflow.facultyId],
-		references: [facultyInWorkflow.id]
+export const attachmentsInWorkflowRelations = relations(attachmentsInWorkflow, ({one}) => ({
+	programme: one(programmes, {
+		fields: [attachmentsInWorkflow.programmeId],
+		references: [programmes.id]
 	}),
-}));
-
-export const facultyInWorkflowRelations = relations(facultyInWorkflow, ({many}) => ({
-	departmentsInWorkflows: many(departmentsInWorkflow),
-}));
-
-export const artifactsInWorkflowRelations = relations(artifactsInWorkflow, ({one}) => ({
+	user: one(users, {
+		fields: [attachmentsInWorkflow.createdBy],
+		references: [users.id]
+	}),
 	processInstancesInWorkflow: one(processInstancesInWorkflow, {
-		fields: [artifactsInWorkflow.processId],
+		fields: [attachmentsInWorkflow.processId],
 		references: [processInstancesInWorkflow.id]
 	}),
 	taskInstancesInWorkflow: one(taskInstancesInWorkflow, {
-		fields: [artifactsInWorkflow.taskId],
+		fields: [attachmentsInWorkflow.taskId],
 		references: [taskInstancesInWorkflow.id]
-	}),
-	programmesInWorkflow: one(programmesInWorkflow, {
-		fields: [artifactsInWorkflow.programmeId],
-		references: [programmesInWorkflow.id]
-	}),
-	usersInWorkflow: one(usersInWorkflow, {
-		fields: [artifactsInWorkflow.createdBy],
-		references: [usersInWorkflow.id]
 	}),
 }));
 
 export const processInstancesInWorkflowRelations = relations(processInstancesInWorkflow, ({one, many}) => ({
-	artifactsInWorkflows: many(artifactsInWorkflow),
+	attachmentsInWorkflows: many(attachmentsInWorkflow),
 	definitionVersionsInWorkflow: one(definitionVersionsInWorkflow, {
 		fields: [processInstancesInWorkflow.definitionVersionId],
 		references: [definitionVersionsInWorkflow.id]
 	}),
-	programmesInWorkflow: one(programmesInWorkflow, {
+	programme: one(programmes, {
 		fields: [processInstancesInWorkflow.programmeId],
-		references: [programmesInWorkflow.id]
+		references: [programmes.id]
 	}),
-	usersInWorkflow: one(usersInWorkflow, {
+	user: one(users, {
 		fields: [processInstancesInWorkflow.startedBy],
-		references: [usersInWorkflow.id]
+		references: [users.id]
 	}),
 	taskInstancesInWorkflows: many(taskInstancesInWorkflow),
 	auditEventsInWorkflows: many(auditEventsInWorkflow),
 }));
 
 export const taskInstancesInWorkflowRelations = relations(taskInstancesInWorkflow, ({one, many}) => ({
-	artifactsInWorkflows: many(artifactsInWorkflow),
+	attachmentsInWorkflows: many(attachmentsInWorkflow),
 	processInstancesInWorkflow: one(processInstancesInWorkflow, {
 		fields: [taskInstancesInWorkflow.processId],
 		references: [processInstancesInWorkflow.id]
@@ -175,23 +100,15 @@ export const taskInstancesInWorkflowRelations = relations(taskInstancesInWorkflo
 	taskInstancesInWorkflows: many(taskInstancesInWorkflow, {
 		relationName: "taskInstancesInWorkflow_causedByTaskId_taskInstancesInWorkflow_id"
 	}),
-	programmesInWorkflow: one(programmesInWorkflow, {
+	programme: one(programmes, {
 		fields: [taskInstancesInWorkflow.programmeId],
-		references: [programmesInWorkflow.id]
+		references: [programmes.id]
 	}),
-	usersInWorkflow: one(usersInWorkflow, {
+	user: one(users, {
 		fields: [taskInstancesInWorkflow.completedBy],
-		references: [usersInWorkflow.id]
+		references: [users.id]
 	}),
 	auditEventsInWorkflows: many(auditEventsInWorkflow),
-}));
-
-export const definitionsInWorkflowRelations = relations(definitionsInWorkflow, ({one, many}) => ({
-	usersInWorkflow: one(usersInWorkflow, {
-		fields: [definitionsInWorkflow.createdBy],
-		references: [usersInWorkflow.id]
-	}),
-	definitionVersionsInWorkflows: many(definitionVersionsInWorkflow),
 }));
 
 export const definitionVersionsInWorkflowRelations = relations(definitionVersionsInWorkflow, ({one, many}) => ({
@@ -199,11 +116,19 @@ export const definitionVersionsInWorkflowRelations = relations(definitionVersion
 		fields: [definitionVersionsInWorkflow.definitionId],
 		references: [definitionsInWorkflow.id]
 	}),
-	usersInWorkflow: one(usersInWorkflow, {
+	user: one(users, {
 		fields: [definitionVersionsInWorkflow.createdBy],
-		references: [usersInWorkflow.id]
+		references: [users.id]
 	}),
 	processInstancesInWorkflows: many(processInstancesInWorkflow),
+}));
+
+export const definitionsInWorkflowRelations = relations(definitionsInWorkflow, ({one, many}) => ({
+	definitionVersionsInWorkflows: many(definitionVersionsInWorkflow),
+	user: one(users, {
+		fields: [definitionsInWorkflow.createdBy],
+		references: [users.id]
+	}),
 }));
 
 export const auditEventsInWorkflowRelations = relations(auditEventsInWorkflow, ({one}) => ({
@@ -215,75 +140,12 @@ export const auditEventsInWorkflowRelations = relations(auditEventsInWorkflow, (
 		fields: [auditEventsInWorkflow.taskId],
 		references: [taskInstancesInWorkflow.id]
 	}),
-	programmesInWorkflow: one(programmesInWorkflow, {
+	programme: one(programmes, {
 		fields: [auditEventsInWorkflow.programmeId],
-		references: [programmesInWorkflow.id]
+		references: [programmes.id]
 	}),
-	usersInWorkflow: one(usersInWorkflow, {
+	user: one(users, {
 		fields: [auditEventsInWorkflow.actorId],
-		references: [usersInWorkflow.id]
+		references: [users.id]
 	}),
-}));
-
-export const phaseStepsInWorkflowRelations = relations(phaseStepsInWorkflow, ({one, many}) => ({
-	phasesInWorkflow: one(phasesInWorkflow, {
-		fields: [phaseStepsInWorkflow.phaseId],
-		references: [phasesInWorkflow.id]
-	}),
-	programmePhaseStepsInWorkflows: many(programmePhaseStepsInWorkflow),
-}));
-
-export const phasesInWorkflowRelations = relations(phasesInWorkflow, ({many}) => ({
-	phaseStepsInWorkflows: many(phaseStepsInWorkflow),
-	programmePhasesInWorkflows: many(programmePhasesInWorkflow),
-}));
-
-export const programmePhasesInWorkflowRelations = relations(programmePhasesInWorkflow, ({one, many}) => ({
-	programmesInWorkflow: one(programmesInWorkflow, {
-		fields: [programmePhasesInWorkflow.programmeId],
-		references: [programmesInWorkflow.id]
-	}),
-	phasesInWorkflow: one(phasesInWorkflow, {
-		fields: [programmePhasesInWorkflow.phaseId],
-		references: [phasesInWorkflow.id]
-	}),
-	programmePhaseStepsInWorkflows: many(programmePhaseStepsInWorkflow),
-}));
-
-export const programmePhaseStepsInWorkflowRelations = relations(programmePhaseStepsInWorkflow, ({one, many}) => ({
-	phaseStepsInWorkflow: one(phaseStepsInWorkflow, {
-		fields: [programmePhaseStepsInWorkflow.phaseStepId],
-		references: [phaseStepsInWorkflow.id]
-	}),
-	programmePhasesInWorkflow: one(programmePhasesInWorkflow, {
-		fields: [programmePhaseStepsInWorkflow.programmePhaseId],
-		references: [programmePhasesInWorkflow.id]
-	}),
-	attachmentsInWorkflows: many(attachmentsInWorkflow),
-}));
-
-export const attachmentsInWorkflowRelations = relations(attachmentsInWorkflow, ({one}) => ({
-	programmePhaseStepsInWorkflow: one(programmePhaseStepsInWorkflow, {
-		fields: [attachmentsInWorkflow.programmePhaseStepId],
-		references: [programmePhaseStepsInWorkflow.id]
-	}),
-	usersInWorkflow: one(usersInWorkflow, {
-		fields: [attachmentsInWorkflow.uploadedBy],
-		references: [usersInWorkflow.id]
-	}),
-}));
-
-export const notificationRecipientsInWorkflowRelations = relations(notificationRecipientsInWorkflow, ({one}) => ({
-	notificationsInWorkflow: one(notificationsInWorkflow, {
-		fields: [notificationRecipientsInWorkflow.notificationId],
-		references: [notificationsInWorkflow.id]
-	}),
-	usersInWorkflow: one(usersInWorkflow, {
-		fields: [notificationRecipientsInWorkflow.recipientId],
-		references: [usersInWorkflow.id]
-	}),
-}));
-
-export const notificationsInWorkflowRelations = relations(notificationsInWorkflow, ({many}) => ({
-	notificationRecipientsInWorkflows: many(notificationRecipientsInWorkflow),
 }));
